+++ sdk/cosmos/azure_data_cosmos/src/constants.rs
@@ -36,6 +36,11 @@ pub const POST_TRIGGER_INCLUDE: HeaderName =
 pub const SESSION_TOKEN: HeaderName = HeaderName::from_static("x-ms-session-token");
 pub const INDEXING_DIRECTIVE: HeaderName = HeaderName::from_static("x-ms-indexing-directive");
 pub const SUB_STATUS: HeaderName = HeaderName::from_static("x-ms-substatus");
+pub const THROUGHPUT_BUCKET: HeaderName = HeaderName::from_static("x-ms-cosmos-throughput-bucket");
+pub const PRIORITY_LEVEL: HeaderName = HeaderName::from_static("x-ms-cosmos-priority-level");
+// Used to allow tentative writes in multi-write accounts.
+pub const ALLOW_TENTATIVE_WRITES: HeaderName =
+    HeaderName::from_static("x-ms-cosmos-allow-tentative-writes");
 
 pub const QUERY_CONTENT_TYPE: ContentType = ContentType::from_static("application/query+json");