+++ sdk/cosmos/azure_data_cosmos/src/cosmos_request.rs
@@ -103,6 +103,18 @@ impl CosmosRequest {
         self.operation_type.is_read_only()
     }
 
+    pub fn client_headers<T: AsHeaders>(&mut self, headers: &T) {
+        // Collect all headers exposed by the `AsHeaders` implementation for client options
+        // always prioritize existing headers in the request over client-level headers.
+        if let Ok(iter) = headers.as_headers() {
+            for (name, value) in iter {
+                if self.headers.get_optional_str(&name).is_none() {
+                    self.headers.insert(name, value);
+                }
+            }
+        }
+    }
+
     /// Gets the corresponding http method for the given `OperationType`.
     pub fn http_method(&self) -> Method {
         self.operation_type.http_method()
@@ -197,8 +209,8 @@ impl CosmosRequestBuilder {
         self
     }
 
-    pub fn headers<T: AsHeaders>(mut self, headers: &T) -> Self {
-        // Collect all headers exposed by the `AsHeaders` implementation.
+    pub fn request_headers<T: AsHeaders>(mut self, headers: &T) -> Self {
+        // Collect all headers exposed by the `AsHeaders` implementation for request options
         // If conversion fails we silently ignore (the caller can always add
         // individual headers via `header()`).
         if let Ok(iter) = headers.as_headers() {
@@ -263,7 +275,9 @@ mod tests {
     use super::*;
     use crate::operation_context::OperationType;
     use crate::resource_context::ResourceType;
-    use crate::{constants, PartitionKey};
+    use crate::{
+        constants, ConsistencyLevel, CosmosClientOptions, ItemOptions, PartitionKey, PriorityLevel,
+    };
 
     fn make_base_request(op: OperationType) -> CosmosRequest {
         let req = CosmosRequest::builder(op, ResourceLink::root(ResourceType::Documents))
@@ -368,6 +382,67 @@ mod tests {
             .iter()
             .any(|(n, _)| n == &constants::IS_UPSERT);
         assert!(has_upsert);
+    }
+
+    #[test]
+    fn prioritize_request_headers_over_client_headers() {
+        let mut request_custom_headers = std::collections::HashMap::new();
+        request_custom_headers.insert(
+            HeaderName::from_static("x-custom-header"),
+            HeaderValue::from_static("custom_value"),
+        );
+
+        let item_options = ItemOptions {
+            consistency_level: Some(ConsistencyLevel::Session),
+            throughput_bucket: Some(1),
+            priority: Some(PriorityLevel::Low),
+            custom_headers: request_custom_headers,
+            ..Default::default()
+        };
+        let req = CosmosRequest::builder(
+            OperationType::Create,
+            ResourceLink::root(ResourceType::Documents),
+        )
+        .request_headers(&item_options)
+        .build()
+        .unwrap();
+
+        let mut req_with_client_headers = req.clone();
+        req_with_client_headers
+            .request_context
+            .location_endpoint_to_route = Some("https://example.com/".parse().unwrap());
+
+        let mut client_custom_headers = std::collections::HashMap::new();
+        client_custom_headers.insert(
+            HeaderName::from_static("x-custom-header"),
+            HeaderValue::from_static("custom_value-2"),
+        );
+
+        let client_options = CosmosClientOptions {
+            consistency_level: Some(ConsistencyLevel::Strong),
+            throughput_bucket: Some(5),
+            priority: Some(PriorityLevel::High),
+            custom_headers: client_custom_headers,
+            ..Default::default()
+        };
+        req_with_client_headers.client_headers(&client_options);
+
+        let raw = req_with_client_headers.into_raw_request();
+        let get_header = |name: &HeaderName| {
+            raw.headers()
+                .iter()
+                .find(|(n, _)| n == &name)
+                .map(|(_, v)| v.as_str())
+                .unwrap()
+        };
+
+        assert_eq!(get_header(&constants::THROUGHPUT_BUCKET), "1");
+        assert_eq!(get_header(&constants::PRIORITY_LEVEL), "Low");
+        assert_eq!(get_header(&constants::CONSISTENCY_LEVEL), "Session");
+        assert_eq!(
+            get_header(&HeaderName::from_static("x-custom-header")),
+            "custom_value"
+        );
     }
 
     #[test]